Fix number rounding
Former-commit-id: 5ae731ea8080f3b0b315292fa479177adafb8577

+++ backend/src/main/kotlin/com/torusage/database/entity/archive/ArchiveGeoRelay.kt
@@ -1,6 +1,7 @@
 package com.torusage.database.entity.archive
 
 import org.torproject.descriptor.NetworkStatusEntry
+import java.math.BigDecimal
 import java.text.SimpleDateFormat
 import java.util.*
 import javax.persistence.EmbeddedId
@@ -14,8 +15,8 @@ import javax.persistence.Entity
 class ArchiveGeoRelay(
     networkStatusEntry: NetworkStatusEntry,
     seenAt: Date,
-    var latitude: Double,
-    var longitude: Double,
+    var latitude: BigDecimal,
+    var longitude: BigDecimal,
 ) {
     @EmbeddedId
     var id: ArchiveNodeId = ArchiveNodeId(
